+++ src/Decorator/EventDecorator/CouponDecorator.php
@@ -10,20 +10,8 @@ use App\Decorator\Events\Coupon;
  * Class Events
  * @package App\Decorator\EventDecorator
  */
-final class CouponDecorator implements EventDecorator
+final class CouponDecorator extends AbstractEventDecorator
 {
-    /** @var EventDecorator */
-    private $eventDecorator;
-
-    /**
-     * Events constructor.
-     * @param EventDecorator $eventDecorator
-     */
-    public function __construct(EventDecorator $eventDecorator)
-    {
-        $this->eventDecorator = $eventDecorator;
-    }
-
     /**
      * @param array $events
      * @return array
