+++ src/Decorator/PriceDecorator/PriceDecoratorInterface.php
@@ -5,17 +5,14 @@ declare(strict_types=1);
 namespace App\Decorator\PriceDecorator;
 
 /**
- * Class OriginalPriceDecorator
+ * Interface PriceDecoratorInterface
  * @package App\Decorator\PriceDecorator
  */
-final class OriginalPriceDecorator implements DecoratorInterface
+interface PriceDecoratorInterface
 {
     /**
      * @param int $totalPrice
      * @return int
      */
-    public function getPrice(int $totalPrice): int
-    {
-        return $totalPrice;
-    }
+    public function getPrice(int $totalPrice): int;
 }
